# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * StreamFlix API - Netflix/Amazon style streaming app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface AuthInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  email: string;
  isAdmin: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type ItemType = typeof ItemType[keyof typeof ItemType];


export const ItemType = {
  movie: 'movie',
  series: 'series',
  documentary: 'documentary',
} as const;

export interface Item {
  id: number;
  title: string;
  description: string;
  genres: string[];
  tags: string[];
  rating: number;
  releaseYear: number;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  backdropUrl?: string | null;
  /** @nullable */
  trailerUrl?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  director?: string | null;
  cast?: string[];
  type?: ItemType;
  /** @nullable */
  interactionCount?: number | null;
  createdAt: string;
}

export type ItemInputType = typeof ItemInputType[keyof typeof ItemInputType];


export const ItemInputType = {
  movie: 'movie',
  series: 'series',
  documentary: 'documentary',
} as const;

export interface ItemInput {
  title: string;
  description: string;
  genres: string[];
  tags: string[];
  rating: number;
  releaseYear: number;
  posterUrl?: string;
  backdropUrl?: string;
  trailerUrl?: string;
  duration?: number;
  director?: string;
  cast?: string[];
  type: ItemInputType;
}

export type ItemUpdateType = typeof ItemUpdateType[keyof typeof ItemUpdateType];


export const ItemUpdateType = {
  movie: 'movie',
  series: 'series',
  documentary: 'documentary',
} as const;

export interface ItemUpdate {
  title?: string;
  description?: string;
  genres?: string[];
  tags?: string[];
  rating?: number;
  releaseYear?: number;
  posterUrl?: string;
  backdropUrl?: string;
  trailerUrl?: string;
  duration?: number;
  director?: string;
  cast?: string[];
  type?: ItemUpdateType;
}

export interface ItemsPage {
  items: Item[];
  total: number;
  page: number;
  limit: number;
}

export type RecommendationRowReason = typeof RecommendationRowReason[keyof typeof RecommendationRowReason];


export const RecommendationRowReason = {
  trending: 'trending',
  new_releases: 'new_releases',
  because_you_watched: 'because_you_watched',
  top_picks: 'top_picks',
  genre_row: 'genre_row',
  content_based: 'content_based',
} as const;

export interface RecommendationRow {
  title: string;
  reason: RecommendationRowReason;
  items: Item[];
  /** @nullable */
  sourceItemTitle?: string | null;
}

export type InteractionInputEventType = typeof InteractionInputEventType[keyof typeof InteractionInputEventType];


export const InteractionInputEventType = {
  view: 'view',
  click: 'click',
  purchase: 'purchase',
  watch: 'watch',
  rate: 'rate',
} as const;

export interface InteractionInput {
  itemId: number;
  eventType: InteractionInputEventType;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  watchDuration?: number | null;
}

export interface InteractionWithItem {
  id: number;
  itemId: number;
  eventType: string;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  watchDuration?: number | null;
  createdAt: string;
  item: Item;
}

export interface WatchlistStatus {
  inWatchlist: boolean;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 10
     */
  rating: number;
  comment?: string;
}

export interface Review {
  id: number;
  itemId: number;
  userId: number;
  /** @nullable */
  userEmail?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface GenreCount {
  genre: string;
  count: number;
}

export interface InteractionTypeCount {
  eventType: string;
  count: number;
}

export interface RecentActivityItem {
  eventType: string;
  itemTitle: string;
  /** @nullable */
  userEmail?: string | null;
  createdAt: string;
}

export interface Analytics {
  totalUsers: number;
  totalItems: number;
  totalInteractions: number;
  topItems: Item[];
  genreDistribution: GenreCount[];
  interactionsByType: InteractionTypeCount[];
  recentActivity: RecentActivityItem[];
}

export interface AdminStats {
  totalUsers: number;
  totalItems: number;
  totalInteractions: number;
  totalWatchlistEntries: number;
  totalReviews: number;
}

export type ListItemsParams = {
genre?: string;
page?: number;
limit?: number;
};

export type GetTrendingParams = {
limit?: number;
genre?: string;
};

export type GetNewReleasesParams = {
limit?: number;
};

export type SearchItemsParams = {
q: string;
genre?: string;
minRating?: number;
limit?: number;
};

export type SimilarItemsParams = {
limit?: number;
};

export type GetInteractionHistoryParams = {
limit?: number;
};