# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamFlix API - Netflix/Amazon style streaming app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── ITEMS ──────────────────────────────────────────────────────────────────
  /items:
    get:
      operationId: listItems
      tags: [items]
      summary: List catalog items with optional filtering
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsPage"
  /items/trending:
    get:
      operationId: getTrending
      tags: [items]
      summary: Get trending items by recent interaction velocity
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: genre
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Trending items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
  /items/new-releases:
    get:
      operationId: getNewReleases
      tags: [items]
      summary: Get recently added items
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: New releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
...
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get a single item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Admin-only item CRUD
  /admin/items:
    post:
      operationId: createItem
      tags: [admin]
      summary: Create a catalog item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /admin/items/{id}:
    patch:
      operationId: updateItem
      tags: [admin]
      summary: Update a catalog item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [admin]
      summary: Delete a catalog item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── SEARCH ─────────────────────────────────────────────────────────────────
...
  /recommendations/similar/{itemId}:
    get:
      operationId: similarItems
      tags: [recommendations]
      summary: Get items similar to a given item (content-based)
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Similar items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
...
  /interactions:
    post:
      operationId: logInteraction
      tags: [interactions]
      summary: Log a user interaction (view, click, rate, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionInput"
      responses:
        "201":
          description: Interaction logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /interactions/history:
    get:
      operationId: getInteractionHistory
      tags: [interactions]
      summary: Get user's viewing/interaction history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Interaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InteractionWithItem"

  # ─── WATCHLIST ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── Auth ─────────────────────────────────────────────────────────────────
    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, createdAt, isAdmin]
      properties:
        id:
          type: integer
        email:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ── Items ─────────────────────────────────────────────────────────────────
    Item:
      type: object
      required: [id, title, description, genres, tags, rating, releaseYear, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        releaseYear:
          type: integer
        posterUrl:
          type: ["string", "null"]
        backdropUrl:
          type: ["string", "null"]
        trailerUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        director:
          type: ["string", "null"]
        cast:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [movie, series, documentary]
        interactionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ItemInput:
      type: object
      required: [title, description, genres, tags, rating, releaseYear, type]
      properties:
        title:
          type: string
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        releaseYear:
          type: integer
        posterUrl:
          type: string
        backdropUrl:
          type: string
        trailerUrl:
          type: string
        duration:
          type: integer
        director:
          type: string
        cast:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [movie, series, documentary]
    ItemUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        genres:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        rating:
          type: number
        releaseYear:
          type: integer
        posterUrl:
          type: string
        backdropUrl:
          type: string
        trailerUrl:
          type: string
        duration:
          type: integer
        director:
          type: string
        cast:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [movie, series, documentary]
    ItemsPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Recommendations ───────────────────────────────────────────────────────
...
    InteractionInput:
      type: object
      required: [itemId, eventType]
      properties:
        itemId:
          type: integer
        eventType:
          type: string
          enum: [view, click, purchase, watch, rate]
        rating:
          type: ["number", "null"]
        watchDuration:
          type: ["integer", "null"]
    InteractionWithItem:
      type: object
      required: [id, itemId, eventType, createdAt, item]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        eventType:
          type: string
        rating:
          type: ["number", "null"]
        watchDuration:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        item:
          $ref: "#/components/schemas/Item"

    # ── Watchlist ─────────────────────────────────────────────────────────────
...